Rename loggers to make memory impl clear
